pyproject.toml
[project]
name = "monster-word-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.21.0",
    "google-cloud-storage>=3.7.0",
    "google-cloud-texttospeech>=2.33.0",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
]
